+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Logs
 
 ### 2017-01-07 v1.0.1
-- Add outerBorder option for Cells
-- Add pagerBgColor option for Swiper
-- Fix Cascade bug in X5, MIUI Browser
+- Add **outerBorder** option for **Cells**
+- Add **pagerBgColor** option for **Swiper**
+- Fix **Cascade** bug in X5, MIUI Browser
+
+### 2017-01-08 v1.0.2
+- Add **ajaxUrl** and **ajaxData** options for **Cascade**, in purpose of loading data by Ajax.
+- Add **fromTemplateUrl** method for **$popup**, in purpose of loading popup template by Ajax.
+- Fix **$storage.on** bug.